+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Custom columns modal style improvements (#671)
 -   Add new ESLint rules to ensure code quality (#684)
+-   Use max rank and round from perf endpoint stats (#670)
 
 ### Fixed